feat(M): adjustments
