+++ frontend/src/app/[locale]/(auth)/login/components/PasswordInput.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 import { EyeIcon, EyeOffIcon, CheckIcon, XIcon } from 'lucide-react';
 import { cn } from '@/lib/utils';
