+++ frontend/src/app/[locale]/(auth)/login/components/PrivacyDialog.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useMemo } from 'react';
 import { Download, Trash2, Server, Lock } from 'lucide-react';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/Card';
