+++ frontend/src/app/[locale]/(main)/feedback/[id]/components/FeedbackDetail.tsx
@@ -19,15 +19,17 @@ import {
 } from '@/components/ui/Accordion';
 import { FeedbackResponse } from '@/interfaces/FeedbackQuoteProps';
 import { useCallback, useEffect, useState } from 'react';
-import { useTranslations } from 'next-intl';
+import { useLocale, useTranslations } from 'next-intl';
 import { getSessionFeedback } from '@/services/client/SessionService';
 import { showErrorToast } from '@/lib/toast';
+import { convertTimeToMinutes, formattedDate } from '@/lib/utils';
 import FeedbackQuote from './FeedbackQuote';
 import FeedbackDialog from './FeedbackDialog';
 import FeedbackDetailLoadingPage from '../loading';
 
 export default function FeedbackDetail({ sessionId }: { sessionId: string }) {
   const t = useTranslations('Feedback');
+  const locale = useLocale();
   const [feedbackDetail, setFeedbackDetail] = useState<FeedbackResponse | null>(null);
   const [isLoading, setIsLoading] = useState(true);
   const getFeedbackDetail = useCallback(
@@ -59,23 +61,6 @@ export default function FeedbackDetail({ sessionId }: { sessionId: string }) {
   useEffect(() => {
     getFeedbackDetail(sessionId);
   }, [getFeedbackDetail, sessionId]);
-
-  const convertTimeToMinutes = (seconds: number) => {
-    return `${Math.floor(seconds / 60)}:${String(seconds % 60).padStart(2, '0')}`;
-  };
-
-  const formatDateTime = (dateString: string) => {
-    const date = new Date(dateString);
-
-    const day = String(date.getDate()).padStart(2, '0');
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const year = date.getFullYear();
-
-    const hours = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-
-    return `${day}.${month}.${year}, ${hours}:${minutes}`;
-  };
 
   const progressBarData = [
     { key: t('progressBars.structure'), value: feedbackDetail?.feedback?.scores.structure ?? 0 },
@@ -126,7 +111,7 @@ export default function FeedbackDetail({ sessionId }: { sessionId: string }) {
       <div className="h-20 bg-marigold-10 px-4 py-5 rounded-md text-center w-full">
         <div className="text-lg text-marigold-90">{feedbackDetail?.title}</div>
         <div className="text-base text-marigold-95">
-          {formatDateTime(feedbackDetail?.createdAt ?? '')}
+          {formattedDate(feedbackDetail?.createdAt, locale)}
         </div>
       </div>
       <FeedbackDialog sessionId={sessionId} />
