+++ frontend/src/app/[locale]/(main)/new-conversation-scenario/components/CategoryButton.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { cn } from '@/lib/utils';
 import { ButtonHTMLAttributes } from 'react';
