+++ frontend/src/lib/utils.ts
@@ -72,10 +72,16 @@ export function PrimaryGoals(): UserOption[] {
   ];
 }
 
-export const formattedDate = (date: string, locale: string) => {
+export const formattedDate = (date: string | undefined, locale: string) => {
+  if (!date) return '';
+
   return new Date(date).toLocaleDateString(locale, {
     day: '2-digit',
     month: '2-digit',
     year: 'numeric',
   });
+};
+
+export const convertTimeToMinutes = (seconds: number) => {
+  return `${Math.floor(seconds / 60)}:${String(seconds % 60).padStart(2, '0')}`;
 };
